ai-engine/requirements.txt
paho-mqtt==1.6.1      # Communication avec ton broker
pandas==2.0.3         # Manipulation des données (tableaux)
scikit-learn==1.3.0   # L'intelligence (Régression Linéaire)
numpy==1.24.3         # Calculs mathématiques
joblib==1.3.1         # Pour sauvegarder/charger ton IA sur le disque
python-dotenv==1.0.0  # Pour tes secrets (IP, mots de passe)